--- packages/chat/src/__fixtures__/conversations.ts
@@ -1,481 +1,0 @@
-/* eslint-disable no-secrets/no-secrets */
-export const MOCK_CONVERSATION_1 = {
-  turns: [
-    {
-      id: 'cm2nrikq700013b6ggzjzqvbm',
-      type: 'system',
-      timestamp: 1729802113136,
-      messages: [
-        {
-          type: 'text',
-          text: [
-            {
-              children: [
-                {
-                  text: "Let's hit the beach! ",
-                },
-              ],
-            },
-            {
-              children: [
-                {
-                  children: [
-                    {
-                      text: '',
-                    },
-                  ],
-                },
-              ],
-            },
-            {
-              children: [
-                {
-                  text: 'Which beach are you thinking?',
-                },
-              ],
-            },
-          ],
-          delay: 1000,
-        },
-      ],
-    },
-    {
-      id: 'cm2nrutnu00013b6gschbhq7y',
-      type: 'user',
-      message: 'asdasd',
-      timestamp: 1729802684586,
-    },
-    {
-      id: 'cm2nruu1100023b6gpzz9a0wk',
-      type: 'system',
-      timestamp: 1729802685061,
-      messages: [
-        {
-          type: 'text',
-          text: [
-            {
-              children: [
-                {
-                  text: "Didn't catch which beach you wanted to go to.",
-                },
-              ],
-            },
-          ],
-          delay: 1000,
-        },
-      ],
-      actions: [
-        {
-          name: 'Marie Curtis',
-          request: {
-            type: 'path-v23i13ew2',
-            payload: {
-              label: 'Marie Curtis',
-              actions: [],
-            },
-          },
-        },
-        {
-          name: 'Sunnyside',
-          request: {
-            type: 'path-kn3pl3ew2',
-            payload: {
-              label: 'Sunnyside',
-              actions: [],
-            },
-          },
-        },
-        {
-          name: "Hanlan's",
-          request: {
-            type: 'path-6a4853eu1',
-            payload: {
-              label: "Hanlan's",
-              actions: [],
-            },
-          },
-        },
-        {
-          name: 'Gibraltar',
-          request: {
-            type: 'path-cw6xb3ecn',
-            payload: {
-              label: 'Gibraltar',
-              actions: [],
-            },
-          },
-        },
-        {
-          name: 'Wards',
-          request: {
-            type: 'path-i36y03ess',
-            payload: {
-              label: 'Wards',
-              actions: [],
-            },
-          },
-        },
-        {
-          name: 'Cherry',
-          request: {
-            type: 'path-9w6yp3enl',
-            payload: {
-              label: 'Cherry',
-              actions: [],
-            },
-          },
-        },
-        {
-          name: 'Woodbine',
-          request: {
-            type: 'path-1g70r3eel',
-            payload: {
-              label: 'Woodbine',
-              actions: [],
-            },
-          },
-        },
-        {
-          name: 'Kew Balmy',
-          request: {
-            type: 'path-gj71n3erx',
-            payload: {
-              label: 'Kew Balmy',
-              actions: [],
-            },
-          },
-        },
-        {
-          name: 'Bluffs',
-          request: {
-            type: 'path-er74f3eia',
-            payload: {
-              label: 'Bluffs',
-              actions: [],
-            },
-          },
-        },
-        {
-          name: 'Centre',
-          request: {
-            type: 'path-1g77i3ehm',
-            payload: {
-              label: 'Centre',
-              actions: [],
-            },
-          },
-        },
-      ],
-    },
-    {
-      id: 'cm2ns108b00013b6gmv7za55n',
-      type: 'user',
-      message: 'asdasd',
-      timestamp: 1729802973035,
-    },
-    {
-      id: 'cm2ns10g800023b6g2hpj8dmb',
-      type: 'system',
-      timestamp: 1729802973320,
-      messages: [
-        {
-          type: 'text',
-          text: [
-            {
-              children: [
-                {
-                  text: 'Sorry, I didn’t get that. Please try again.',
-                },
-              ],
-            },
-          ],
-          delay: 1000,
-        },
-      ],
-      actions: [
-        {
-          name: 'Marie Curtis',
-          request: {
-            type: 'path-v23i13ew2',
-            payload: {
-              label: 'Marie Curtis',
-              actions: [],
-            },
-          },
-        },
-        {
-          name: 'Sunnyside',
-          request: {
-            type: 'path-kn3pl3ew2',
-            payload: {
-              label: 'Sunnyside',
-              actions: [],
-            },
-          },
-        },
-        {
-          name: "Hanlan's",
-          request: {
-            type: 'path-6a4853eu1',
-            payload: {
-              label: "Hanlan's",
-              actions: [],
-            },
-          },
-        },
-        {
-          name: 'Gibraltar',
-          request: {
-            type: 'path-cw6xb3ecn',
-            payload: {
-              label: 'Gibraltar',
-              actions: [],
-            },
-          },
-        },
-        {
-          name: 'Wards',
-          request: {
-            type: 'path-i36y03ess',
-            payload: {
-              label: 'Wards',
-              actions: [],
-            },
-          },
-        },
-        {
-          name: 'Cherry',
-          request: {
-            type: 'path-9w6yp3enl',
-            payload: {
-              label: 'Cherry',
-              actions: [],
-            },
-          },
-        },
-        {
-          name: 'Woodbine',
-          request: {
-            type: 'path-1g70r3eel',
-            payload: {
-              label: 'Woodbine',
-              actions: [],
-            },
-          },
-        },
-        {
-          name: 'Kew Balmy',
-          request: {
-            type: 'path-gj71n3erx',
-            payload: {
-              label: 'Kew Balmy',
-              actions: [],
-            },
-          },
-        },
-        {
-          name: 'Bluffs',
-          request: {
-            type: 'path-er74f3eia',
-            payload: {
-              label: 'Bluffs',
-              actions: [],
-            },
-          },
-        },
-        {
-          name: 'Centre',
-          request: {
-            type: 'path-1g77i3ehm',
-            payload: {
-              label: 'Centre',
-              actions: [],
-            },
-          },
-        },
-      ],
-    },
-    {
-      id: 'cm2ns130j00033b6g1fxke4au',
-      type: 'user',
-      message: 'w234324',
-      timestamp: 1729802976643,
-    },
-    {
-      id: 'cm2ns136p00043b6g6y2q00ts',
-      type: 'system',
-      timestamp: 1729802976865,
-      messages: [
-        {
-          type: 'text',
-          text: [
-            {
-              children: [
-                {
-                  text: 'Sorry, I didn’t get that. Please try again.',
-                },
-              ],
-            },
-          ],
-          delay: 1000,
-        },
-      ],
-      actions: [
-        {
-          name: 'Marie Curtis',
-          request: {
-            type: 'path-v23i13ew2',
-            payload: {
-              label: 'Marie Curtis',
-              actions: [],
-            },
-          },
-        },
-        {
-          name: 'Sunnyside',
-          request: {
-            type: 'path-kn3pl3ew2',
-            payload: {
-              label: 'Sunnyside',
-              actions: [],
-            },
-          },
-        },
-        {
-          name: "Hanlan's",
-          request: {
-            type: 'path-6a4853eu1',
-            payload: {
-              label: "Hanlan's",
-              actions: [],
-            },
-          },
-        },
-        {
-          name: 'Gibraltar',
-          request: {
-            type: 'path-cw6xb3ecn',
-            payload: {
-              label: 'Gibraltar',
-              actions: [],
-            },
-          },
-        },
-        {
-          name: 'Wards',
-          request: {
-            type: 'path-i36y03ess',
-            payload: {
-              label: 'Wards',
-              actions: [],
-            },
-          },
-        },
-        {
-          name: 'Cherry',
-          request: {
-            type: 'path-9w6yp3enl',
-            payload: {
-              label: 'Cherry',
-              actions: [],
-            },
-          },
-        },
-        {
-          name: 'Woodbine',
-          request: {
-            type: 'path-1g70r3eel',
-            payload: {
-              label: 'Woodbine',
-              actions: [],
-            },
-          },
-        },
-        {
-          name: 'Kew Balmy',
-          request: {
-            type: 'path-gj71n3erx',
-            payload: {
-              label: 'Kew Balmy',
-              actions: [],
-            },
-          },
-        },
-        {
-          name: 'Bluffs',
-          request: {
-            type: 'path-er74f3eia',
-            payload: {
-              label: 'Bluffs',
-              actions: [],
-            },
-          },
-        },
-        {
-          name: 'Centre',
-          request: {
-            type: 'path-1g77i3ehm',
-            payload: {
-              label: 'Centre',
-              actions: [],
-            },
-          },
-        },
-      ],
-    },
-    {
-      id: 'cm2ns15k800053b6ger346qd5',
-      type: 'user',
-      message: 'Wards',
-      timestamp: 1729802979944,
-    },
-    {
-      id: 'cm2ns1ggw00063b6g38eteq74',
-      type: 'system',
-      timestamp: 1729802994080,
-      messages: [
-        {
-          type: 'text',
-          text: [
-            {
-              children: [
-                {
-                  text: '',
-                },
-                {
-                  text: 'wards',
-                },
-                {
-                  text: ' beach is safe to swim at.',
-                },
-              ],
-            },
-            {
-              children: [
-                {
-                  text: 'It measured ',
-                },
-                {
-                  text: 'null',
-                },
-                {
-                  text: ' ppm of E. coli, ',
-                },
-                {
-                  text: '2 months ago',
-                },
-                {
-                  text: ' ',
-                },
-              ],
-            },
-          ],
-          delay: 1000,
-        },
-        {
-          type: 'END',
-        },
-      ],
-    },
-  ],
-};
