--- packages/chat/src/views/ChatWindow/ChatWindow.css.ts
@@ -1,5 +1,0 @@
-import { style } from '@vanilla-extract/css';
-
-export const chatContent = style({
-  padding: '0 20px',
-});
